requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.4.0=pypi_0
adjusttext=0.8=pypi_0
appnope=0.1.3=pyhd8ed1ab_0
asttokens=2.2.1=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
bzip2=1.0.8=h620ffc9_4
ca-certificates=2023.5.7=hf0a4a13_0
cachetools=5.3.1=pypi_0
certifi=2023.5.7=pypi_0
charset-normalizer=3.1.0=pypi_0
classeval=0.2.0=pypi_0
clusteval=2.2.1=pypi_0
clustimage=1.5.16=pypi_0
colourmap=1.1.14=pypi_0
contourpy=1.1.0=pypi_0
cycler=0.11.0=pypi_0
cython=0.29.35=pypi_0
datazets=0.1.6=pypi_0
debugpy=1.5.1=py311h313beb8_0
decorator=5.1.1=pyhd8ed1ab_0
distfit=1.6.10=pypi_0
executing=1.2.0=pyhd8ed1ab_0
flatbuffers=23.5.26=pypi_0
fonttools=4.40.0=pypi_0
funcsigs=1.0.2=pypi_0
gast=0.4.0=pypi_0
google-auth=2.20.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.56.0=pypi_0
h5py=3.9.0=pypi_0
hdbscan=0.8.29=pypi_0
idna=3.4=pypi_0
imagehash=4.3.1=pypi_0
imageio=2.31.1=pypi_0
importlib-metadata=6.7.0=pyha770c72_0
importlib_metadata=6.7.0=hd8ed1ab_0
ipykernel=6.15.0=pyh736e0ef_0
ipython=8.14.0=pyhd1c38e8_0
ismember=1.0.2=pypi_0
jedi=0.18.2=pyhd8ed1ab_0
joblib=1.2.0=pypi_0
jupyter_client=8.3.0=pyhd8ed1ab_0
jupyter_core=5.3.0=py311hca03da5_0
keras=2.13.1rc0=pypi_0
kiwisolver=1.4.4=pypi_0
lazy-loader=0.2=pypi_0
libclang=16.0.0=pypi_0
libcxx=16.0.6=h4653b0c_0
libffi=3.4.4=hca03da5_0
libsodium=1.0.18=h27ca646_1
markdown=3.4.3=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=3.1=pypi_0
numpy=1.25.0=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.7.0.72=pypi_0
openssl=3.1.1=h53f4e23_1
opt-einsum=3.3.0=pypi_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.2=pypi_0
parso=0.8.3=pyhd8ed1ab_0
patsy=0.5.3=pypi_0
pca=2.0.3=pypi_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.5.0=pypi_0
pip=23.1.2=py311hca03da5_0
platformdirs=3.8.0=pyhd8ed1ab_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
protobuf=4.23.3=pypi_0
psutil=5.9.0=py311h80987f9_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pydot=1.4.2=pypi_0
pygments=2.15.1=pyhd8ed1ab_0
pyparsing=3.1.0=pypi_0
pypickle=1.1.0=pypi_0
python=3.11.3=hb885b13_1
python-dateutil=2.8.2=pyhd8ed1ab_0
pytz=2023.3=pypi_0
pywavelets=1.4.1=pypi_0
pyzmq=25.1.0=py311h313beb8_0
readline=8.2=h1a28f6b_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scatterd=1.3.5=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.11.0=pypi_0
seaborn=0.12.2=pypi_0
setuptools=67.8.0=py311hca03da5_0
six=1.16.0=pyh6c4a22f_0
sklearn=0.0.post5=pypi_0
sqlite=3.41.2=h80987f9_0
stack_data=0.6.2=pyhd8ed1ab_0
statsmodels=0.14.0=pypi_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tensorflow=2.13.0rc1=pypi_0
tensorflow-estimator=2.13.0rc0=pypi_0
tensorflow-macos=2.13.0rc1=pypi_0
termcolor=2.3.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2023.4.12=pypi_0
tk=8.6.12=hb8d0fd4_0
tornado=6.2=py311h80987f9_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
typing-extensions=4.6.3=hd8ed1ab_0
typing_extensions=4.6.3=pyha770c72_0
tzdata=2023.3=pypi_0
urllib3=1.26.16=pypi_0
wcwidth=0.2.6=pyhd8ed1ab_0
werkzeug=2.3.6=pypi_0
wheel=0.38.4=py311hca03da5_0
wrapt=1.15.0=pypi_0
xz=5.4.2=h80987f9_0
zeromq=4.3.4=hbdafb3b_1
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h5a0b063_0
